pyproject.toml
[tool.poetry]
name = "django-xrpl-test"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"
packages = [{include = "django_xrp_test"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
pytest = "^7.2.0"
isort = "^5.11.4"
black = "^22.12.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
django = "^4.1.5"
pytest = "^7.2.0"
isort = "^5.11.4"
black = "^22.12.0"
django-environ = "^0.9.0"
django-debug-toolbar = "^3.8.1"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.5"
xrpl-py = "^1.7.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
djangorestframework-simplejwt = "^5.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
